rename all resource.py to http.py
reason: resource.py is an old name that should not be used
