+++ wazo_phoned/plugins/api/plugin.py
@@ -1,7 +1,7 @@
 # Copyright 2019 The Wazo Authors  (see the AUTHORS file)
 # SPDX-License-Identifier: GPL-3.0-or-later
 
-from .resource import SwaggerResource
+from .http import SwaggerResource
 
 
 class Plugin:
