+++ wazo_phoned/plugins/cisco/http.py
@@ -1,7 +1,7 @@
 # Copyright 2019 The Wazo Authors  (see the AUTHORS file)
 # SPDX-License-Identifier: GPL-3.0-or-later
 
-from wazo_phoned.plugin_helpers.proxy.resource import ProxyMenu, ProxyInput, ProxyLookup
+from wazo_phoned.plugin_helpers.proxy.http import ProxyMenu, ProxyInput, ProxyLookup
 
 
 class Menu(ProxyMenu):
