+++ wazo_phoned/plugins/snom/http.py
@@ -1,7 +1,11 @@
 # Copyright 2019 The Wazo Authors  (see the AUTHORS file)
 # SPDX-License-Identifier: GPL-3.0-or-later
 
-from wazo_phoned.plugin_helpers.proxy.resource import ProxyLookup
+from wazo_phoned.plugin_helpers.proxy.http import ProxyInput, ProxyLookup
+
+
+class Input(ProxyInput):
+    pass
 
 
 class Lookup(ProxyLookup):
