+++ wazo_phoned/plugins/snom/plugin.py
@@ -2,7 +2,7 @@
 # SPDX-License-Identifier: GPL-3.0-or-later
 
 from wazo_phoned.plugin_helpers.proxy.plugin import ProxyPlugin
-from .resource import Input, Lookup
+from .http import Input, Lookup
 
 
 class Plugin(ProxyPlugin):
