+++ wazo_phoned/plugins/status/plugin.py
@@ -1,7 +1,7 @@
 # Copyright 2019 The Wazo Authors  (see the AUTHORS file)
 # SPDX-License-Identifier: GPL-3.0-or-later
 
-from .resource import Status
+from .http import Status
 
 
 class Plugin:
