+++ wazo_phoned/plugins/thomson/http.py
@@ -1,11 +1,7 @@
 # Copyright 2019 The Wazo Authors  (see the AUTHORS file)
 # SPDX-License-Identifier: GPL-3.0-or-later
 
-from wazo_phoned.plugin_helpers.proxy.resource import ProxyInput, ProxyLookup
-
-
-class Input(ProxyInput):
-    pass
+from wazo_phoned.plugin_helpers.proxy.http import ProxyLookup
 
 
 class Lookup(ProxyLookup):
